Move stuff around in resto
