+++ app/src/main/java/be/ugent/zeus/hydra/resto/extrafood/Food.java
@@ -1,4 +1,4 @@
-package be.ugent.zeus.hydra.resto;
+package be.ugent.zeus.hydra.resto.extrafood;
 
 import android.os.Parcel;
 import android.os.Parcelable;
