+++ app/src/main/java/be/ugent/zeus/hydra/resto/extrafood/FoodAdapter.java
@@ -5,7 +5,6 @@ import android.view.ViewGroup;
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.common.ui.ViewUtils;
 import be.ugent.zeus.hydra.common.ui.recyclerview.adapters.ItemDiffAdapter;
-import be.ugent.zeus.hydra.resto.Food;
 
 /**
  * Adapter for food items.
