+++ app/src/main/java/be/ugent/zeus/hydra/resto/history/DayRequest.java
@@ -1,0 +1,10 @@
+package be.ugent.zeus.hydra.resto.history;
+
+/**
+ * Request the menu for a single day.
+ *
+ * @author Niko Strijbol
+ */
+public class DayRequest {
+
+}
