+++ app/src/main/java/be/ugent/zeus/hydra/resto/menu/RestoFragment.java
@@ -28,9 +28,8 @@ import be.ugent.zeus.hydra.common.arch.observers.SuccessObserver;
 import be.ugent.zeus.hydra.MainActivity;
 import be.ugent.zeus.hydra.common.ui.BaseActivity;
 import be.ugent.zeus.hydra.resto.RestoMenu;
-import be.ugent.zeus.hydra.resto.network.SelectableMetaRequest;
 import be.ugent.zeus.hydra.resto.RestoPreferenceFragment;
-import be.ugent.zeus.hydra.resto.locations.RestoLocationActivity;
+import be.ugent.zeus.hydra.resto.meta.RestoLocationActivity;
 import be.ugent.zeus.hydra.resto.sandwich.SandwichActivity;
 import be.ugent.zeus.hydra.resto.extrafood.ExtraFoodActivity;
 import be.ugent.zeus.hydra.utils.Analytics;
