+++ app/src/main/java/be/ugent/zeus/hydra/resto/menu/SelectableMetaViewModel.java
@@ -4,7 +4,6 @@ import android.app.Application;
 
 import be.ugent.zeus.hydra.common.request.Request;
 import be.ugent.zeus.hydra.common.ui.RequestViewModel;
-import be.ugent.zeus.hydra.resto.network.SelectableMetaRequest;
 
 import java.util.List;
