+++ app/src/main/java/be/ugent/zeus/hydra/resto/meta/MetaRequest.java
@@ -1,8 +1,7 @@
-package be.ugent.zeus.hydra.resto.network;
+package be.ugent.zeus.hydra.resto.meta;
 
 import android.support.annotation.NonNull;
 
-import be.ugent.zeus.hydra.resto.RestoMeta;
 import be.ugent.zeus.hydra.common.network.Endpoints;
 import be.ugent.zeus.hydra.common.network.JsonSpringRequest;
 import be.ugent.zeus.hydra.common.caching.Cache;
