+++ app/src/main/java/be/ugent/zeus/hydra/resto/meta/Resto.java
@@ -1,4 +1,4 @@
-package be.ugent.zeus.hydra.resto;
+package be.ugent.zeus.hydra.resto.meta;
 
 import android.os.Parcel;
 import android.os.Parcelable;
