+++ app/src/main/java/be/ugent/zeus/hydra/resto/meta/RestoLocationActivity.java
@@ -1,4 +1,4 @@
-package be.ugent.zeus.hydra.resto.locations;
+package be.ugent.zeus.hydra.resto.meta;
 
 import android.Manifest;
 import android.arch.lifecycle.ViewModelProviders;
@@ -19,8 +19,6 @@ import be.ugent.zeus.hydra.common.arch.observers.ErrorObserver;
 import be.ugent.zeus.hydra.common.arch.observers.ProgressObserver;
 import be.ugent.zeus.hydra.common.arch.observers.SuccessObserver;
 import be.ugent.zeus.hydra.common.ui.BaseActivity;
-import be.ugent.zeus.hydra.resto.Resto;
-import be.ugent.zeus.hydra.resto.RestoMeta;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.MapFragment;
