+++ app/src/main/java/be/ugent/zeus/hydra/resto/sandwich/SandwichAdapter.java
@@ -5,7 +5,6 @@ import android.view.ViewGroup;
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.common.ui.ViewUtils;
 import be.ugent.zeus.hydra.common.ui.recyclerview.adapters.MultiSelectDiffAdapter;
-import be.ugent.zeus.hydra.resto.Sandwich;
 
 /**
  * Adapter for sandwiches.
