+++ app/src/main/java/be/ugent/zeus/hydra/resto/sandwich/SandwichHolder.java
@@ -8,7 +8,6 @@ import android.widget.TextView;
 import be.ugent.zeus.hydra.R;
 import be.ugent.zeus.hydra.common.ui.recyclerview.adapters.MultiSelectDiffAdapter;
 import be.ugent.zeus.hydra.common.ui.recyclerview.viewholders.DataViewHolder;
-import be.ugent.zeus.hydra.resto.Sandwich;
 import net.cachapa.expandablelayout.ExpandableLayout;
 
 /**
