+++ app/src/main/java/be/ugent/zeus/hydra/resto/sandwich/SandwichViewModel.java
@@ -5,8 +5,6 @@ import android.app.Application;
 import be.ugent.zeus.hydra.common.request.Request;
 import be.ugent.zeus.hydra.common.request.Requests;
 import be.ugent.zeus.hydra.common.ui.RequestViewModel;
-import be.ugent.zeus.hydra.resto.Sandwich;
-import be.ugent.zeus.hydra.resto.network.SandwichRequest;
 
 import java.util.Arrays;
 import java.util.List;
