+++ app/src/main/res/layout/activity_resto_location.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".resto.locations.RestoLocationActivity" >
+    tools:context=".resto.meta.RestoLocationActivity" >
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
