+++ app/src/test/java/be/ugent/zeus/hydra/resto/extrafood/FoodTest.java
@@ -1,6 +1,7 @@
-package be.ugent.zeus.hydra.resto;
+package be.ugent.zeus.hydra.resto.extrafood;
 
 import be.ugent.zeus.hydra.common.ModelTest;
+import be.ugent.zeus.hydra.resto.extrafood.Food;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 import org.junit.Test;
@@ -8,16 +9,17 @@ import org.junit.Test;
 /**
  * @author Niko Strijbol
  */
-public class RestoMetaTest extends ModelTest<RestoMeta> {
+public class FoodTest extends ModelTest<Food> {
 
-    public RestoMetaTest() {
-        super(RestoMeta.class);
+    public FoodTest() {
+        super(Food.class);
     }
 
     @Test
     public void equalsAndHash() {
-        EqualsVerifier.forClass(RestoMeta.class)
+        EqualsVerifier.forClass(Food.class)
                 .suppress(Warning.NONFINAL_FIELDS)
                 .verify();
+
     }
 }
