+++ app/src/test/java/be/ugent/zeus/hydra/resto/menu/MenuFilterTest.java
@@ -1,10 +1,12 @@
-package be.ugent.zeus.hydra.resto;
+package be.ugent.zeus.hydra.resto.menu;
 
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 
 import be.ugent.zeus.hydra.BuildConfig;
 import be.ugent.zeus.hydra.TestApp;
+import be.ugent.zeus.hydra.resto.RestoMenu;
+import be.ugent.zeus.hydra.resto.RestoPreferenceFragment;
 import be.ugent.zeus.hydra.testing.Utils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
