+++ app/src/test/java/be/ugent/zeus/hydra/resto/menu/MenuRequestTest.java
@@ -1,4 +1,4 @@
-package be.ugent.zeus.hydra.resto.network;
+package be.ugent.zeus.hydra.resto.menu;
 
 import android.content.SharedPreferences;
 import android.os.Build;
