+++ app/src/test/java/be/ugent/zeus/hydra/resto/meta/RestoMetaTest.java
@@ -1,6 +1,7 @@
-package be.ugent.zeus.hydra.resto;
+package be.ugent.zeus.hydra.resto.meta;
 
 import be.ugent.zeus.hydra.common.ModelTest;
+import be.ugent.zeus.hydra.resto.meta.RestoMeta;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 import org.junit.Test;
@@ -8,17 +9,16 @@ import org.junit.Test;
 /**
  * @author Niko Strijbol
  */
-public class FoodTest extends ModelTest<Food> {
+public class RestoMetaTest extends ModelTest<RestoMeta> {
 
-    public FoodTest() {
-        super(Food.class);
+    public RestoMetaTest() {
+        super(RestoMeta.class);
     }
 
     @Test
     public void equalsAndHash() {
-        EqualsVerifier.forClass(Food.class)
+        EqualsVerifier.forClass(RestoMeta.class)
                 .suppress(Warning.NONFINAL_FIELDS)
                 .verify();
-
     }
 }
